requirements.txt
asgiref
crispy-bootstrap5
distlib
Django
django-cors-headers
django-crispy-forms
django-environ
django-filter
djangorestframework
EasyProcess
entrypoint2
filelock
Markdown
MouseInfo
mss
Pillow
platformdirs
PyAutoGUI
PyGetWindow
PyMsgBox
pyperclip
PyRect
pyscreenshot
PyScreeze
pytweening
pytz
sqlparse
tzdata
virtualenv
